app/utils/classification/requirements.txt
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
requests==2.28.1
torch==1.9.0+cu102
torchvision==0.10.0+cu102
